% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbouretro-package.R
\docType{package}
\name{bbouretro-package}
\alias{bbouretro}
\alias{bbouretro-package}
\title{bbouretro: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID})

Other contributors:
\itemize{
  \item Seb Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID}) [contributor]
  \item Colton Stephens \email{colton@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-8744-6405}{ORCID}) [contributor]
  \item John Boulanger \email{boulange@ecological.bc.ca} [contributor]
}

}
\keyword{internal}
