% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec-to-cc.R
\name{bbr_cc_to_rec}
\alias{bbr_cc_to_rec}
\title{Calf Cow Ratio to Recruitment}
\usage{
bbr_cc_to_rec(x, sex_ratio = 0.5)
}
\arguments{
\item{x}{A numeric vector of the calf:cow ratio}

\item{sex_ratio}{A}
}
\value{
A numeric vector of the equivalent recruitment rate
}
\description{
The calf cow ratios is simply the number of calves divided by the number of cows.
As described by DeCesare et al. (2012) in order to convert the
calf cow ratio to the female recruitment rate it is necessary to
multiple the calf cow ratio by the sex ratio to get the
female calf to cow ratio and then divide that number by itself plus 1 to
get the female recruitment rate ie female calves divided by all females.
To perform the inverse conversion see \code{\link[=bbr_rec_to_cc]{bbr_rec_to_cc()}}
}
\examples{
bbr_cc_to_rec(c(0, 1, 0.5, NA))
}
\seealso{
\code{\link[=bbr_rec_to_cc]{bbr_rec_to_cc()}}
}
