% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km-survival.R
\name{bbr_km_survival}
\alias{bbr_km_survival}
\title{Calculate survival}
\format{
A tibble with columns:
\describe{
\item{PopulationName}{Population name}
\item{Year}{Year sampled}
\item{S}{Survival estimate }
\item{S_SE}{SE}
\item{S_CIL}{Confidence limit}
\item{S_CIU}{Confidence limit}
\item{MeanMonitored}{Mean number of caribou monitored each month}
\item{sumdead }{Total number of mortalities in a year}
\item{sumalive}{Total number of caribou-months in a year}
\item{Status}{Indicates less than 12 months monitored or if there were 0 mortalities in a given year}
}
}
\usage{
bbr_km_survival(x, mort_type = "total", variance = "pollock")
}
\arguments{
\item{x}{input survival data frame}

\item{mort_type}{Mortality data to be included.   Values can be “total” which
would be all (MortalitiesCertain and MortalitiesUncertain) or “certain”,
(MortalitiesCertain only).}

\item{variance}{Variance type to estimate.  Can be the Greenwood estimator
“greenwood” or Pollock estimator (“pollock”).}
}
\value{
An output data frame with the columns.
}
\description{
This function estimates survival rates based on the Kaplan Meir survival rate
estimator (Pollock et al. 1989).
}
\details{
See the vignette Methods for description of equations used.
}
\examples{
survival_est <- bbr_km_survival(
  bboudata::bbousurv_a,
  mort_type = "total",
  variance = "greenwood"
)
}
\references{
Pollock, K. H., S. R. Winterstein, C. M. Bunck, and P. D. Curtis.
1989. Survival analysis in telemetry studies: the staggered entry design.
Journal of Wildlife Management 53:7-15. TODO Need Cox paper
}
