% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{bbr_lambda}
\alias{bbr_lambda}
\title{Simulate lambda}
\usage{
bbr_lambda(recruitment, survival)
}
\arguments{
\item{recruitment}{A data frame generated by \code{bbr_recruitment()}.}

\item{survival}{A data frame generated by \code{bbr_survival()}.}
}
\value{
A data.frame.
}
\description{
This function uses the output of \code{bbr_survival()} and \code{bbr_recruitment()}
to estimate \eqn{\lambda} using the Hatter-Bergerud equation (Hatter and
Bergerud, 1991). Monte Carlo simulation is used to generate confidence limits
on \eqn{\lambda}.
}
\details{
See the \code{vignette("methods", package = "bbouretro")} for
descriptions of the equations used. The raw_values can be plotted using
\code{bbr_plot_lambda_distributions()} and the summary data frame can be output
using \code{bbr_lambda_summarize()} or plotted using \code{bbr_plot_lambda()}.
}
\examples{
\dontrun{
bbr_lambda(recruitment_est, survival_est)
}
}
\references{
Hatter, Ian, and Wendy Bergerud. 1991. “Moose Recruitment, Adult
Mortality and Rate of Change” 27: 65–73.
}
