% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-simulate.R
\name{bbr_lambda_simulate}
\alias{bbr_lambda_simulate}
\title{Simulate lambda}
\usage{
bbr_lambda_simulate(recruitment, survival)
}
\arguments{
\item{recruitment}{A data frame generated by \code{bbr_recruitment()}.}

\item{survival}{A data frame generated by \code{bbr_km_survival()}.}
}
\value{
A list that contains two data frames. First one is named raw_values
can contains raw simulated data and the second one is the summary data.
}
\description{
This function uses the output of \code{bbr_km_survival()} and \code{bbr_recruitment()}
to estimate \eqn{\lambda} using the Hatter-Bergerud equation. Monte Carlo
simulation is used to generate confidence limits on \eqn{\lambda}.
}
\details{
See the \code{vignette("Methods", package = "bbouretro")} for
descriptions of the equations used. The raw_values can be plotted using
\code{bbr_plot_lambda_distributions()} and the summary data frame can be output
using \code{bbr_lambda_summarize()} or plotted using \code{bbr_plot_lamba()}.
}
\examples{
\dontrun{
bbr_lambda_simulate(recruitment_est, survival_est)
}
}
