% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-summarize.R
\name{bbr_lambda_summarize}
\alias{bbr_lambda_summarize}
\title{Summarize lambda}
\format{
The return object has these columns:
\describe{
\item{PopulationName}{Population name}
\item{Year}{Year sampled}
\item{S}{Estimated survival}
\item{R}{Estimated recruitment}
\item{Lambda}{Estimated lambda}
\item{Lambda_SE }{SE}
\item{Lambda_CIL}{Percentile 95\% confidence limits}
\item{Lambda_CIU }{Percentile 95\% confidence limits}
\item{prop_lgt1 }{Proportion simulations where lambda>1}
\item{mean_sim_survival}{Mean simulated survival value}
\item{mean_sim_recruitment}{Mean simulated recruitment value}
\item{mean_sim_lambda }{Mean simulated lambda value}
\item{median_sim_lambda}{Median simulated lambda value}
}
}
\usage{
bbr_lambda_summarize(lambda)
}
\arguments{
\item{lambda}{A list object generated by \code{bbr_lambda_simulate()}.}
}
\value{
A data frame. The columns are listed in the format section.
}
\description{
Provides a summary of yearly \eqn{\lambda} estimates from simulations.
}
\examples{
\dontrun{
bbr_lambda_summarize(lambda)
}
}
