% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-growth-distributions.R
\name{bbr_plot_growth_distributions}
\alias{bbr_plot_growth_distributions}
\title{Plot population growth distributions}
\usage{
bbr_plot_growth_distributions(growth)
}
\arguments{
\item{growth}{A data frame generated by \code{bbr_growth()}.}
}
\value{
A ggplot object.
}
\description{
Create histograms of simulated population growth (\eqn{\lambda}) values.
}
\details{
Plots are generated that show the distribution of simulated
population growth (\eqn{\lambda}) values, the mean estimate (red line).  In
addition, a hashed line indicates where \eqn{\lambda = 1}.  Plots allow
users to evaluate the symmetry of the distributions of \eqn{\lambda}.
}
\examples{
\dontrun{
recruitment_est <- bbr_recruitment(bboudata::bbourecruit_a)
survival_est <- bbr_survival(bboudata::bbousurv_a)
growth_est <- bbr_growth(survival_est, recruitment_est)

bbr_plot_growth_distributions(growth_est)
}
}
