% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-lambda.R
\name{bbr_plot_lambda}
\alias{bbr_plot_lambda}
\title{Plot lambda}
\usage{
bbr_plot_lambda(lambda)
}
\arguments{
\item{lambda}{Summary output object generated by \code{bbr_lambda_simulate()}.}
}
\value{
A ggplot object.
}
\description{
Uses output from \code{bbr_lambda_simulate()}.
}
\details{
A plot of \eqn{lambda} estimates is given for each population unit.
}
\examples{
\dontrun{
bbr_plot_lambda(lambda_est)
}
}
