% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-recruitment.R
\name{bbr_plot_recruitment}
\alias{bbr_plot_recruitment}
\title{Plot recruitment}
\usage{
bbr_plot_recruitment(recruitment)
}
\arguments{
\item{recruitment}{Input summary data frame generated by \code{bbr_recruitment()}.}
}
\value{
A ggplot object.
}
\description{
Uses output from \code{bbr_recruitment()} to plot yearly estimates.
}
\details{
A plot of yearly survival is given for each population unit.
}
\examples{
\dontrun{
bbr_plot_recruitment(recruitment_est)
}
}
