% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec-to-cc.R
\name{bbr_rec_to_cc}
\alias{bbr_rec_to_cc}
\title{Recruitment to Calf Cow Ratio}
\usage{
bbr_rec_to_cc(x, sex_ratio = 0.5)
}
\arguments{
\item{x}{A numeric vector of the recruitment rate}

\item{sex_ratio}{A}
}
\value{
A numeric vector of the equivalent calf:cow ratio
}
\description{
Converts the female recruitment rate to the calf cow ratio.
For further information see \code{\link[=bbr_cc_to_rec]{bbr_cc_to_rec()}}.
}
\examples{
bbr_rec_to_cc(c(0, 1, 0.5, NA))
}
\seealso{
\code{\link[=bbr_cc_to_rec]{bbr_cc_to_rec()}}
}
