% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{bbr_survival}
\alias{bbr_survival}
\title{Estimate survival}
\format{
The return object has these columns:
\describe{
\item{PopulationName}{Population name}
\item{Year}{Year sampled}
\item{estimate}{Survival estimate }
\item{se}{SE}
\item{lower}{Confidence limit}
\item{upper}{Confidence limit}
\item{mean_monitored}{Mean number of caribou monitored each month}
\item{sum_dead }{Total number of mortalities in a year}
\item{sum_alive}{Total number of caribou-months in a year}
\item{status}{Indicates less than 12 months monitored or if there were 0 mortalities in a given year}
}
}
\usage{
bbr_survival(
  x,
  include_uncertain_morts = FALSE,
  variance = "greenwood",
  year_start = 4L
)
}
\arguments{
\item{x}{A data frame that has survival data.}

\item{include_uncertain_morts}{A flag indicating whether to include uncertain mortalities in total mortalities.
The default value is FALSE.}

\item{variance}{Variance type to estimate. Can be the Greenwood estimator
\code{"greenwood"} or Cox Oakes estimator \code{"cox_oakes"}. The default is
"greenwood".}

\item{year_start}{A whole number between 1 and 12 indicating the month of the
start of the caribou (i.e., biological) year. By default, April is set as
the start of the caribou year.}
}
\value{
A data frame. The columns are listed in the format section.
}
\description{
Estimate survival rates based on the Kaplan-Meier survival rate estimator
(Pollock et al. 1989).
}
\details{
\code{x} needs to be formatted in a certain manner. To confirm the input
data frame is in the right format you can use the
\href{https://poissonconsulting.github.io/bboudata/reference/bbd_chk_data_survival.html}{\code{bbd_chk_data_survival}}
function. See the \code{vignette("methods", package = "bbouretro")} for the
equations used in this function.
}
\examples{
survival_est <- bbr_survival(
  bboudata::bbousurv_a,
  include_uncertain_morts = TRUE,
  variance = "greenwood"
)
survival_est <- bbr_survival(
  bboudata::bbousurv_b,
  include_uncertain_morts = FALSE,
  variance = "cox_oakes"
)
}
\references{
Pollock, K. H., S. R. Winterstein, C. M. Bunck, and P. D. Curtis.
1989. Survival analysis in telemetry studies: the staggered entry design.
Journal of Wildlife Management 53:7-15.
}
