% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{KMsurvival}
\alias{KMsurvival}
\title{Calculate survival}
\usage{
KMsurvival(Ssamps, MortType, variance)
}
\arguments{
\item{Ssamps}{}

\item{MortType}{}

\item{variance}{}
}
\description{
Calculate survival
}
