% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-sim.R
\name{lambda_sim}
\alias{lambda_sim}
\title{Simulate lambda}
\usage{
lambda_sim(recruitment, survival)
}
\arguments{
\item{recruitment}{Input summary data frame generated by \code{recruitment()}}

\item{survival}{Input summary data frame generated by \code{km_survival()}}
}
\value{
A list object that contains raw simulated data and summary data.  The
raw data can be plotted using the \code{plot_lambda_distributions()} and the
summary data can be output using the \code{summarize_lambda()} function or
plotted using the \code{plot_lamba()} function.
}
\description{
This function uses the summary output of \code{km_survival()} and \code{recruitment()}
to estimate \eqn{lambda} using the Hatter-Bergerud equation.  Monte Carlo
simulation is used to generate confidence limits on \eqn{lambda}.
}
\details{
The basic Hatter-Bergerud equation used to estimate lamba is given
below. \eqn{lambda = (S) / (1 - R)} The standard errors on estimates of
survival and recruitment are then used to create simulated distributions of
S and R and derive a distribution of λ based on the simulated values.  It
is assumed both survival and recruitment are distributed on the logit scale
therefore bounding each value between 0 and 1.  One thousand simulations
are conducted. Percentile-based confidence limits are then derived.  Also,
the proportion of simulations where \eqn{lambda} is greater than 1 is
tabulated to provide a p-value for the hypothesis test \eqn{(Ho: lambda >=
  1, Ha: lambda < 1)}.  Output summaries also include mean values simulated
for \eqn{S}, \eqn{R}, as well as mean and median \eqn{lambda} values from
the simulation.  Users can plot the distribution of simulated \eqn{lambda}
values using \code{plot_lambda_distributions()} function as well as estimated
values using the \code{plot_lambda()} function.
}
\examples{
\dontrun{
lambda_sim(recruitment_estimate, survival_est)
}
}
