% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-lambda.R
\name{plot_lambda}
\alias{plot_lambda}
\title{Plot lambda}
\usage{
plot_lambda(lambda)
}
\arguments{
\item{lambda}{Summary output object generated by LambdaSim}
}
\value{
A ggplot object
}
\description{
Uses output from LambdaSim
}
\details{
A plot of \eqn{lambda} estimates is given for each population unit
}
\examples{
\dontrun{
plot_lambda(lambda_est)
}
}
