% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-lambda-dstr.R
\name{plot_lambda_distributions}
\alias{plot_lambda_distributions}
\title{Plot lambda distributions}
\usage{
plot_lambda_distributions(lambda, population)
}
\arguments{
\item{lambda}{Summary output object generated by \code{lambda_sim()}}

\item{population}{Name of a population to select.}
}
\value{
A ggplot object
}
\description{
Provides histograms of simulated \eqn{lambda} values for a chose population
unit.
}
\details{
Plots are generated that show the distribution of simulated λ
values, the mean estimate (red line).  In addition, a hashed line indicates
where \eqn{lambda = 1}.  Plots allow users to evaluate the symmetry of the distributions
of \eqn{lambda}.
}
\examples{
\dontrun{
plot_lambda_distributions(lambda)
}
}
