% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-recruitment.R
\name{plot_recruitment}
\alias{plot_recruitment}
\title{Plot recruitment}
\usage{
plot_recruitment(recruitment)
}
\arguments{
\item{recruitment}{Input summary data frame generated by Recruitment}
}
\value{
A ggplot object
}
\description{
Uses output from Recruitment to plot yearly estimates
}
\details{
A plot of yearly survival is given for each population unit.
}
\examples{
\dontrun{
plot_recruitment(recruitment_est)
}
}
