% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-survival.R
\name{plot_survival}
\alias{plot_survival}
\title{Plot survival}
\usage{
plot_survival(survival)
}
\arguments{
\item{survival}{Input summary data frame generated by \code{km_survival()}}
}
\value{
A ggplot object
}
\description{
Uses output from \code{km_survival()} to plot yearly estimates.
}
\details{
A plot of yearly survival is given for each population unit.
}
\examples{
\dontrun{
plot_survival(survival_est)
}
}
