% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-lambda-dstr.R
\name{plotLambdaDistributions}
\alias{plotLambdaDistributions}
\title{Plot lambda distributions}
\usage{
plotLambdaDistributions(L, Pop)
}
\arguments{
\item{L}{}

\item{Pop}{}
}
\description{
Plot lambda distributions
}
