% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{recruitment}
\alias{recruitment}
\title{Calculate recruitment}
\format{
A tibble with columns:
\describe{
\item{PopulationName}{Population name}
\item{Year}{Year sampled}
\item{R}{Recruitment estimate }
\item{R_SE}{SE}
\item{R_CIL}{Confidence limit}
\item{R_CIU}{Confidence limit}
\item{groups}{Groups sampled}
\item{FemaleCalves}{Estimated female calves}
\item{Females}{Estimated adult females}
\item{sexratio}{Input sex ratio}
\item{pFemales}{Input proportion adult females}
}
}
\usage{
recruitment(x, pFemales, sexratio, variance)
}
\arguments{
\item{x}{input recruitment data frame}

\item{pFemales}{Assumed or estimated proportion females in the population
used to assign unknown sex caribou (see details).  Can be set to 0 to
exclude unknown sex caribou from recruitment estimates}

\item{sexratio}{Sex ratio of caribou at birth used to assign calves and
yearlings as male or female (see details).  Sex ratio is defined as the
proportion females at birth.  Usually this is set at 0.5.}

\item{variance}{Estimate variance using “binomial” or “bootstrap”}
}
\value{
An output data frame with the columns.
}
\description{
This function generates estimates of recruitment from the Recruitment data
frame with confidence limits.   User’s can input the assumed proportion of
females in the population (to estimate females from adult caribou that have
unknown sex) as well as sex ratio at birth. Variance can be estimated using a
few different approaches.
}
\details{
See the vignette Methods for description of equations used.
}
\examples{
recruitment_estimate <- recruitment(
  bboudata::bbourecruit_a, 
  pFemales = 0.65, 
  sexratio = 0.5, 
 variance <- "binomial"
)
}
\references{
DeCesare, N. J., M. Hebblewhite, M. Bradley, K. G. Smith, D.
Hervieux, and L. Neufeld. 2012. Estimating ungulate recruitment and growth
rates using age ratios. The Journal of Wildlife Management 76:144-153.
Pearson, A., J. Boulanger, and J. L. Thorley. 2022. Boreal Caribou Monitoring
– Literature Review of Current and Historical Data Collection and Analysis
Methods Used to Estimate Survival, Recruitment, and Population Growth.
}
