% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-sum.R
\name{summarize_lambda}
\alias{summarize_lambda}
\title{Summarize lambda}
\format{
A tibble with columns:
\describe{
\item{PopulationName}{Population name}
\item{Year}{Year sampled}
\item{S}{Estimated survival}
\item{R}{Estimated recruitment}
\item{Lambda}{Estimated lambda (S/(1-R))}
\item{SE_Lambda }{SE}
\item{Lambda_LCL}{Percentile 95\% confidence limits}
\item{Lambda_UCL }{Percentile 95\% confidence limits}
\item{Prop_LGT1 }{Proportion simulations where lambda>1}
\item{meanSimSurv}{Mean simulated survival value}
\item{meanRsim}{Mean simulated recruitment value}
\item{meanSimLambda }{Mean simulated lambda value}
\item{medianSimLambda}{Median simulated lambda value}
}
}
\usage{
summarize_lambda(lambda)
}
\arguments{
\item{lambda}{Summary output object generated by \code{lambda_sim()}}
}
\value{
The following values are given as a data frame.  See LambdaSim for
more details on each row.
}
\description{
Provides a summary of yearly \eqn{lambda} estimates from simulations.
}
\examples{
\dontrun{
summarize_lambda(lambda)
}
}
