% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbouretro-package.R
\docType{package}
\name{bbouretro-package}
\alias{bbouretro}
\alias{bbouretro-package}
\title{bbouretro: Traditional Survival, Recruitment and Population Growth Methods}
\description{
Estimates annual survival, recruitment and population growth using the traditional methods. This package is part of the bbou suite of tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://poissonconsulting.github.io/bbouretro/}
  \item \url{https://github.com/poissonconsulting/bbouretro}
  \item Report bugs at \url{https://github.com/poissonconsulting/bbouretro/issues}
}

}
\author{
\strong{Maintainer}: Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID})

Authors:
\itemize{
  \item John Boulanger \email{boulange@ecological.bc.ca} (\href{https://orcid.org/0000-0001-8222-1445}{ORCID})
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})
}

Other contributors:
\itemize{
  \item Seb Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID}) [contributor]
  \item Colton Stephens \email{colton@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-8744-6405}{ORCID}) [contributor]
  \item Province of Alberta [copyright holder]
}

}
\keyword{internal}
