% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-sim.R
\name{bbr_lambda_simulate}
\alias{bbr_lambda_simulate}
\title{Simulate lambda}
\usage{
bbr_lambda_simulate(recruitment, survival)
}
\arguments{
\item{recruitment}{Input summary data frame generated by \code{bbr_recruitment()}.}

\item{survival}{Input summary data frame generated by \code{bbr_km_survival()}.}
}
\value{
A list object that contains raw simulated data and summary data.  The
raw data can be plotted using the \code{bbr_plot_lambda_distributions()} and the
summary data can be output using the \code{bbr_lambda_summarize()} function or
plotted using the \code{bbr_plot_lamba()} function.
}
\description{
This function uses the summary output of \code{bbr_km_survival()} and
\code{bbr_recruitment()} to estimate \eqn{lambda} using the Hatter-Bergerud
equation. Monte Carlo simulation is used to generate confidence limits on
\eqn{lambda}.
}
\details{
See the vignette Methods for description of equations used. Users
can plot the distribution of simulated \eqn{lambda} values using
\code{bbr_plot_lambda_distributions()} function as well as estimated values
using the \code{bbr_plot_lambda()} function.
}
\examples{
\dontrun{
bbr_lambda_simulate(recruitment_est, survival_est)
}
}
