% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-growth.R
\name{bbr_plot_growth}
\alias{bbr_plot_growth}
\title{Plot population growth}
\usage{
bbr_plot_growth(lambda)
}
\arguments{
\item{lambda}{A data frame generated by \code{bbr_growth()}.}
}
\value{
A ggplot object.
}
\description{
A plot of population growth (\eqn{\lambda}) estimates is given for each
population unit.
}
\examples{
\dontrun{
bbr_plot_growth(lambda_est)
}
}
