% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-lambda.R
\name{bbr_plot_lambda}
\alias{bbr_plot_lambda}
\title{Plot lambda}
\usage{
bbr_plot_lambda(lambda)
}
\arguments{
\item{lambda}{A list object generated by \code{bbr_growth()}.}
}
\value{
A ggplot object.
}
\description{
A plot of \eqn{\lambda} estimates is given for each population unit.
}
\examples{
\dontrun{
bbr_plot_lambda(lambda_est)
}
}
