% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-lambda-distributions.R
\name{bbr_plot_lambda_distributions}
\alias{bbr_plot_lambda_distributions}
\title{Plot lambda distributions}
\usage{
bbr_plot_lambda_distributions(lambda)
}
\arguments{
\item{lambda}{A data frame generated by \code{bbr_lambda()}.}
}
\value{
A ggplot object.
}
\description{
Create histograms of simulated \eqn{\lambda} values for a chosen population
unit.
}
\details{
Plots are generated that show the distribution of simulated
\eqn{\lambda} values, the mean estimate (red line).  In addition, a hashed
line indicates where \eqn{\lambda = 1}.  Plots allow users to evaluate the
symmetry of the distributions of \eqn{\lambda}.
}
\examples{
\dontrun{
bbr_plot_lambda_distributions(lambda)
}
}
