% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-survival.R
\name{bbr_plot_survival}
\alias{bbr_plot_survival}
\title{Plot survival}
\usage{
bbr_plot_survival(survival)
}
\arguments{
\item{survival}{Input summary data frame generated by \code{bbr_km_survival()}.}
}
\value{
A ggplot object.
}
\description{
Uses output from \code{bbr_km_survival()} to plot yearly estimates.
}
\details{
A plot of yearly survival is given for each population unit.
}
\examples{
\dontrun{
bbr_plot_survival(survival_est)
}
}
