% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{bbr_recruitment}
\alias{bbr_recruitment}
\title{Estimate recruitment}
\format{
The return object has these columns:
\describe{
\item{PopulationName}{Population name}
\item{Year}{Year sampled}
\item{R}{Recruitment estimate }
\item{R_SE}{SE}
\item{R_CIL}{Confidence limit}
\item{R_CIU}{Confidence limit}
\item{groups}{Groups sampled}
\item{FemaleCalves}{Estimated female calves}
\item{Females}{Estimated adult females}
\item{sexratio}{Input sex ratio}
\item{p_females}{Input proportion adult females}
}
}
\usage{
bbr_recruitment(x, p_females = 0.65, sexratio = 0.5, variance = "binomial")
}
\arguments{
\item{x}{A data frame that has recruitment data.}

\item{p_females}{Assumed or estimated proportion of females in the population
used to assign unknown sex caribou. Values must be between 0
and 1. Can be set to 0 to exclude unknown sex caribou from recruitment
estimates. The default is set at 0.65.}

\item{sexratio}{Sex ratio of caribou at birth used to assign calves and
yearlings as male or female.  Sex ratio is defined as the
proportion females at birth. Values must be between 0 and 1. The default is
set at 0.5.}

\item{variance}{Estimate variance using "binomial" or "bootstrap". The
default is set as "binomial".}
}
\value{
A data frame. The columns are listed in the format section.
}
\description{
Estimate recruitment using DeCesare et al. (2012) methods.
}
\details{
\code{x} needs to be formatted in a certain manner. To confirm the input
data frame is in the right format you can use the
\link[bboudata:bbd_chk_data_recruitment]{bbd_chk_data_recruitment} function.  See the
\code{vignette("Methods", package = "bbouretro")} for the equations used in
this function.

User’s can input the assumed proportion of females in the population (to
estimate females from adult caribou that have unknown sex) as well as sex
ratio at birth.
}
\examples{
recruitment_est <- bbr_recruitment(
  bboudata::bbourecruit_a,
  p_females = 0.65,
  sexratio = 0.5,
  variance = "binomial"
)
recruitment_est <- bbr_recruitment(
  bboudata::bbourecruit_a,
  p_females = 0.60,
  sexratio = 0.65,
  variance = "bootstrap"
)
}
\references{
DeCesare, Nicholas J., Mark Hebblewhite, Mark Bradley, Kirby G. Smith,
David Hervieux, and Lalenia Neufeld. 2012 “Estimating Ungulate Recruitment
and Growth Rates Using Age Ratios.” The Journal of Wildlife Management
76 (1): 144–53 https://doi.org/10.1002/jwmg.244.
}
