% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km-survival.R
\name{bbr_survival}
\alias{bbr_survival}
\title{Estimate survival}
\format{
The return object has these columns:
\describe{
\item{PopulationName}{Population name}
\item{Year}{Year sampled}
\item{estimate}{Survival estimate }
\item{se}{SE}
\item{lower}{Confidence limit}
\item{upper}{Confidence limit}
\item{mean_monitored}{Mean number of caribou monitored each month}
\item{sum_dead }{Total number of mortalities in a year}
\item{sum_alive}{Total number of caribou-months in a year}
\item{status}{Indicates less than 12 months monitored or if there were 0 mortalities in a given year}
}
}
\usage{
bbr_survival(x, mort_type = "total", variance = "greenwood")
}
\arguments{
\item{x}{A data frame that has survival data.}

\item{mort_type}{Mortality data to be included. Values can be \code{"total"} which
would be all mortalities (MortalitiesCertain and MortalitiesUncertain) or
only \code{"certain"} mortalities (MortalitiesCertain). The default is
\code{"total"}.}

\item{variance}{Variance type to estimate. Can be the Greenwood estimator
\code{"greenwood"} or Cox Oakes estimator \code{"cox_oakes"}. The default is "greenwood".}
}
\value{
A data frame. The columns are listed in the format section.
}
\description{
Estimate survival rates based on the Kaplan-Meier survival rate estimator
(Pollock et al. 1989).
}
\details{
\code{x} needs to be formatted in a certain manner. To confirm the input
data frame is in the right format you can use the
\href{https://poissonconsulting.github.io/bboudata/reference/bbd_chk_data_survival.html}{\code{bbd_chk_data_survival}}
function. See the \code{vignette("Methods", package = "bbouretro")} for the
equations used in this function.
}
\examples{
survival_est <- bbr_survival(
  bboudata::bbousurv_a,
  mort_type = "total",
  variance = "greenwood"
)
survival_est <- bbr_survival(
  bboudata::bbousurv_b,
  mort_type = "certain",
  variance = "cox_oakes"
)
}
\references{
Pollock, K. H., S. R. Winterstein, C. M. Bunck, and P. D. Curtis.
1989. Survival analysis in telemetry studies: the staggered entry design.
Journal of Wildlife Management 53:7-15.
}
