% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{Recruitment}
\alias{Recruitment}
\title{Calculate recruitment}
\usage{
Recruitment(Rsamps, pFemales, sexratio, variance)
}
\arguments{
\item{Rsamps}{}

\item{pFemales}{}

\item{sexratio}{}

\item{variance}{}
}
\description{
Calculate recruitment
}
\examples{
Rsamps<-RsampsC
pFemales<-0.65
sexratio<-0.5
variance<-"binomial"
#... tbd
}
